Move performance dashboard -> Archived page
